Added some features to the roadtrip page like removing a stop, some tools like zooming in on a stop, zooming out to fit all stops, and auto scrolling to a stop's information card. Also added a global to top of page button and integrated the temporary database api data into the browse roadtrips page.

+++ src/app/core/components/Locations/new-location-form/new-location-form.component.ts
@@ -87,8 +87,8 @@ export class NewLocationFormComponent implements OnInit, OnChanges {
     let latVal = this.form.get('latitude')?.value
     let lngVal = this.form.get('longitude')?.value
     this.coordinatesDefined.emit({
-      lat: parseInt(latVal),
-      lng: parseInt(lngVal)
+      lat: parseFloat(latVal),
+      lng: parseFloat(lngVal)
     })
   }
 }
